dont delete last old backups! THX to Marco Lenhardt!

+++ backup-docker-compose.sh
@@ -61,5 +61,9 @@ for i in $COMPOSE; do
 done
 echo -e "\n$TIMESTAMP Backup for Compose Projects completed\n"
 
-find $BACKUPDIR -name "*.tar.gz" -daystart -mtime +$DAYS -delete
-
+# dont delete last old backups!
+OLD_BACKUPS=$(ls -1 $BACKUPDIR/*.gz |wc -l)
+if [ $OLD_BACKUPS -gt 3 ]
+then
+        find $BACKUPDIR -name "*.tar.gz" -daystart -mtime +$DAYS -delete
+fi
